Add constituency items to SentimentDatums (convert from namedtuple to a simple class)
Use the new _asdict method when writing a dataset
Read trees back from the classifier data files.
Add a test of the tree reading by adding trees to one of the test datasets.

Move the fake datasets to test_data.py for better organization - now they are all with the reading test code